requirements.txt
fastapi>=0.115.2
uvicorn>=0.32.0
gradio>=5.9.1
python-dotenv>=1.0.1
scenedetect[opencv]>=0.6.4
opencv-python-headless>=4.10.0
dashscope>=1.20.14
Pillow>=11.0.0
faster-whisper>=1.0.0
